kl_gmres: GMRES linear solver

SIAMFANLEquations.kl_gmres — Function

kl_gmres(x0, b, atv, V, eta, ptv=nothing; orth = "cgs2", side="left", pdata=nothing)

Gmres linear solver. Handles preconditioning and (coming soon) restarts. Uses gmres_base which is completely oblivious to these things.

The deal is

Input:

x0: initial iterate, this is usually zero for nonlinear solvers

b: right hand side (duh!)

atv: matrix-vector product which depends on precomputed data pdta I expect you to use pdata most or all of the time, so it is not an optional argument, even if it's nothing (at least for now). If your mat-vec is just A*v, you have to write a function where A is the precomputed data. API for atv is av=atv(v,pdata)

V: Preallocated n x K array for the Krylov vectors. I store the initial normalized residual in column 1, so you have at most K-1 iterations before gmres_base returns a failure. kl_gmres will handle the restarts.

eta: Termination happens when ||b - Ax|| <= eta || b ||

ptv: preconditioner-vector product, which will also use pdata. The default is nothing, which is no preconditioning at all. API for ptv is px=ptv(x,pdata)

Keyword arguments

pdata: precomputed data. The default is nothing, but that ain't gonna work well for nonlinear equations.

orth: your choice of the wise default, classical Gram-Schmidt twice, or something slower and less stable. Those are classical once (really bad) or a couple variants of modified Gram-Schmidt. mgs2 is what I used in my old matlab codes. Not terrible, but far from great.

side: left or right preconditioning. The default is "left".

Other parameters on the way.

Output:

A named tuple (sol, reshist, lits, idid)

where

sol= final result reshist = residual norm history lits = number of iterations idid = status of the iteration true -> converged false -> failed to converge