nsolsc: scalar equation solver

SIAMFANLEquations.nsolsc — Method

nsolsc(x,f; rtol=1.e-6, atol=1.e-12, maxit=10, fp=difffp, solver="newton", sham=1, armmax=10, keepsolhist=true)

Newton's method for scalar equations. Has most of the features a code for systems of equations needs.

Input:

x: initial iterate

f: function

Options:

fp: derivative. If your derivative function is fp, you give me its name. For example fp=foobar tells me that foobar is your function for the derivative. The default is a forward difference Jacobian that I provide.

Your best bet is to put f and fp in the same file. See the atan example.

solver:

Your choices are Newton or Secant. However, you have sham at your disposal only if you chose Newton.

If you use secant and your initial iterate is poor, you have made a mistake. I will help you by driving the line search with a finite difference derivative.

sham:

This is the Shamanskii method. If sham=1, you have Newton. The iteration updates the derivative every sham iterations. The covergence rate has local q-order sham+1 if you only count iteratons where you update the derivative. You need not provide your own derivative function to use this option.

I only turn Shamanskii on if the residuals are decreasing rapidly, at least a factor of 10, and the line search is quiescent.

rtol, atol: real and absolute error tolerances

maxit: upper bound on number of nonlinear iterations

sham: update Jacobian every sham iteraitons. sham=1 –> Newton

armmax: upper bound on stepsize reductions in linesearch

keepsolhist: if true you get the history of the iteration in the output tuple. This is on by default for scalar equations and off for systems. Only turn it on if you have use for the data, which can get REALLY LARGE.

Output: A tuple (solution, functionval, history) where (iteration counter, f(x), dt)

idid=true if the iteration succeeded and false if not.

solhist=entire history of the iteration if keepsolhist=true