ptcsc: pseuto-transient continuation

SIAMFANLEquations.ptcsc — Method

ptcsc(u, f; rtol=1.e-6, atol=1.e-12, fp=difffp, dt0=1.e-6, maxit=100, keepsolhist=true)

Scalar pseudo-transient continuation solver. PTC is designed to find stable steady state solutions of

du/dt = - f(u)

It is ABSOLUTELY NOT a general purpose nonlinear solver.

Input:

x: initial iterate

f: function

Options:

rtol, atol: real and absolute error tolerances

fp: derivative. If your derivative function is fp, you give me its name. For example fp=foobar tells me that foobar is your function for the derivative. The default is a forward difference Jacobian that I provide.

dt0: initial time step. The default value of 1.e-6 is a bit conservative and is one option you really should play with. Look at the example where I set it to 1.0!

maxit: upper bound on number of nonlinear iterations. This is coupled to dt0. If your choice of dt0 is too small (conservative) then you'll need many iterations to converge and will need a larger value of maxit.

keepsolhist: if true you get the history of the iteration in the output tuple. This is on by default for scalar equations and off for systems. Only turn it on if you have use for the data, which can get REALLY LARGE.

Output: A tuple (solution, functionval, history, idid, solhist) where History is a three column array

(iteration counter, f(x), dt)

idid=true if the iteration succeeded and false if not.

solhist=entire history of the iteration if keepsolhist=true

If the iteration it's time to play with the tolerances, dt0, and maxit. You are certian to fail if there is no solution to the equation.