ptcsol: Pseudo-Transient Continuation Solver

SIAMFANLEquations.ptcsol — Method

ptcsol(F!, x0, FS, FPS, J! = diffjac!; rtol=1.e-6, atol=1.e-12, maxit=20, dt0=1.e-6, dx=1.e-7, pdata = nothing, jfact = klfact, printerr = true, keepsolhist = false)

C. T. Kelley, 2020

Julia versions of the nonlinear solvers from my SIAM books. Herewith: some new stuff ==> ptcsol

You must allocate storage for the function and Jacobian in advance –> in the calling program <– ie. in FS and FPS

Inputs:

F!: function evaluation, the ! indicates that F! overwrites FS, your preallocated storage for the function.

FV=F!(FV,x) or FV=F!(FV,x,pdata) returns FV=F(x)

x0: initial iterate

FS: Preallcoated storage for function. It is an N x 1 column vector

You may dimension it as (n,) or (n,1). (n,) is best, but the
solvers can deal with it either way.

FPS: preallcoated storage for Jacobian. It is an N x N matrix

J!: Jacobian evaluation, the ! indicates that J! overwrites FPS, your preallocated storage for the Jacobian. If you leave this out the default is a finite difference Jacobian.

FP=J!(FP,FV,x) or FP=J!(FP,FV,x,pdata) returns FP=F'(x);
(FP,FV, x) must be the argument list, even if FP does not need FV.
One reason for this is that the finite-difference Jacobian
does and that is the default in the solver.

Lemme tell ya 'bout precision. I designed this code for full precision
functions and linear algebra in any precision you want. You can declare
FPS as Float64, Float32, or Float16 and ptcsol will do the right thing if
YOU do not destroy the declaration in your J! function. I'm amazed
that this works so easily. If the Jacobian is reasonably well 
conditioned, I can see no reason to do linear algebra in 
double precision.

Keyword Arguments (kwargs):

rtol and atol: relative and absolute error tolerances

dt0: initial time step. The default value of 1.e-3 is a bit conservative and is one option you really should play with. Look at the example where I set it to 1.0!

maxit: limit on nonlinear iterations, default=100.

This is coupled to dt0. If your choice of dt0 is too small (conservative) then you'll need many iterations to converge and will need a larger value of maxit

For PTC you'll need more iterations than for a stright-up nonlinear solve. This is part of the price for finding the stable solution.

dx: default = 1.e-7

difference increment in finite-difference derivatives h=dx*norm(x)+1.e-6

pdata:

precomputed data for the function/Jacobian. Things will go better if you use this rather than hide the data in global variables within the module for your function/Jacobian

jfact: default = klfact (tries to figure out best choice)

If your Jacobian has any special structure, please set jfact to the correct choice for a factorization.

I use jfact when I call PTCUpdate to evaluate the Jacobian (using your J!) and factor it. The default is to use klfact (an internal function) to do something reasonable. For general matrices, klfact picks lu! to compute an LU factorization and share storage with the Jacobian. You may change LU to something else by, for example, setting jfact = cholseky! if your Jacobian is spd.

klfact knows about banded matrices and picks qr. You should, however RTFM, allocate the extra two upper bands, and use jfact=qr! to override klfact.

If you give me something that klfact does not know how to dispatch on, then nothing happens. I just return the original Jacobian matrix and ptcsol will use backslash to compute the Newton step.

I know that this is probably not optimal in your situation, so it is good to pick something else, like jfact = lu.

printerr: default = true

I print a helpful message when the solver fails. To supress that message set printerr to false.

keepsolhist: default = false

Set this to true to get the history of the iteration in the output tuple. This is on by default for scalar equations and off for systems. Only turn it on if you have use for the data, which can get REALLY LARGE.

Output:

A named tuple (solution, functionval, history, stats, idid, errcode, solhist) where

solution = converged result functionval = F(solution) history = the vector of residual norms (||F(x)||) for the iteration stats = named tuple of the history of (ifun, ijac, iarm), the number of functions/derivatives/steplength reductions at each iteration.

I do not count the function values for a finite-difference derivative because they count toward a Jacobian evaluation.

idid=true if the iteration succeeded and false if not.

errcode = 0 if if the iteration succeeded = -1 if the initial iterate satisifies the termination criteria = 10 if no convergence after maxit iterations = 1 if the line search failed

solhist:

This is the entire history of the iteration if you've set keepsolhist=true

solhist is an N x K array where N is the length of x and K is the number of iteration + 1. So, for scalar equations, it's a row vector.